safari-600.7-branch/LayoutTests/fullscreen/full-screen-iframe-not-allowed.html

<p>Test for <a href="http://bugs.webkit.org/show_bug.cgi?id=56264">bug 56264</a>: 
Handle entering full screen security restrictions</p>
<p>To test manually, click the "Go full screen" button - the page should not enter full screen mode.</p>
<script src="full-screen-test.js"></script>
<script>
function runTest() {
    var frame = document.getElementById('frame');

    waitForEvent(frame.contentDocument, 'webkitfullscreenchange', function() { 
            consoleWrite("FAIL - entered full screen!"); 
            endTest();
    });

    waitForEventAndEnd(frame.contentDocument, 'webkitfullscreenerror', function() {
        consoleWrite("SUCCEED - did not enter full screen!");
    });

    runWithKeyDown(function() {
        frame.contentDocument.documentElement.webkitRequestFullScreen();
    });
}
</script>
<iframe id="frame" src="resources/inner.html" onload="runTest()">
</iframe>
